Refactor logs prepend/append
