+++ crates/celo-revm/src/common/mod.rs
@@ -1,4 +1,4 @@
 pub mod fee_currency_context;
 pub use fee_currency_context::FeeCurrencyContext;
 
-pub mod global_cip_64_context;
+pub mod global_fee_currency_context;
